Convert spotify Jupyter Notebook to Python script.
Due to the chance that the Spotify API may be accessed for other purposes
or by other Notebooks, it was copied over to a Python script. The original
notebook will be preserved until it's definitely no longer needed.

Script was modified so that the authorization header does not need to be
passed to methods any longer. Rather, it is generated each time it is needed.
this is to prevent the access token from expiring, and to allow calling
the methods without needing to re-generate the header each time.